Move CMAC functionality to a new file
